+++ KoboDeluxe-0.5.1/sound/audio.c
@@ -319,7 +319,7 @@ int timestamp(void)         // IOHAVOC -- remove inline specifier
 {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
-	return tv.tv_sec * 1000000 + tv.tv_usec;
+	return (int)(tv.tv_sec * 1000000 + tv.tv_usec);
 }
 #  else
 inline int timestamp(void)
